requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai
langchain
langchain-openai
langchain-google-genai
PyMuPDF
faiss-cpu
langchain-community
langchain-experimental
numpy
tiktoken